requirements.txt
torch>=2.0.0 
torchvision>=0.15.0 
grad-cam>=1.4.8 
scikit-learn>=1.3.0 
xgboost>=2.0.0 
shap>=0.44.0 
pandas>=2.0.0 
numpy>=1.24.0 
matplotlib>=3.7.0 
seaborn>=0.12.0 
fastapi>=0.104.0 
uvicorn>=0.24.0 
python-multipart>=0.0.6 
pillow>=10.0.0 
joblib>=1.3.0 